dhcp/native_ms - fixes #1054 - Native MSDHCP fails to initialize sometimes
This error is due to timeouts I think. This is because the implementation
of the options lookup is poor. It calls the ms dhcp server for each host
in the reservation list for each maanaged scope. This is very, very, very,
vey slow. This fix dumps the server's reservation list, processes it once
and then uses the cached values for building the options lists. This reduced
the web inteface wait from 36 seconds to ~2 seconds for my installation.