fixes #4332 - initial abrt proxy support
This change makes it possible for the proxy to pretend it is ABRT
(Automatic Bug Reporting Tool[1]) server. When it receives a problem
report it stores it to the disk, and optionally also forwards the
original report to an actual ABRT server.

Forwarding the stored reports to the Foreman server is done by
smart-proxy-abrt-send script which is invoked from cron. It reads the
reports, groups them by the host they came from, optionally merges those
with the same error (the satyr gem and library are needed), and sends
them to the Foreman, one report per host, masquerading as a report from
Puppet.

Non strictly ABRT-related changes:
- Move module Proxy::ChefProxy to Proxy::Request
- Add :foreman_ssl_ options to settings example

[1] https://github.com/abrt/abrt/wiki/ABRT-Project