fixes #13237 - Use RRData option when deleting DNS records
To be able to delete dns records with `dnscmd` without specifying
the `RRData` option, the account needs `Full Control` privileges on the
zone.

`Full Control` allows the account to do much more besides
adding/removing records and this represents a security risk.

Instead of trying to delete all records with a single dnscmd, use
`dnscmd /EnumRecords` and then delete each record individually.  In most
cases, I'd expect `/EnumRecords` to only return a single record anyway
(but I have tested it with many).

There is one side-effect of specifying the IP address when deleting an A
record.  If the matching PTR also exists on the same DNS server, I
believe it gets automatically deleted too.  This hasn't caused any
issues in testing.

Signed-off-by: Alexander Fisher <alex@linfratech.co.uk>