Update register-service.rb
The question asking what user to run as was confusing for us as it defaults on the screen to "default user/logged in user", and was presented in a manner making a user think it's a yes or no question.  There are numerous other reasons why the windows service might fail to get created.  We thought for the longest time it was because the service creation file (this file), didnt allow a domain\username to be entered.  It wasn't until we looked at the file that we realized its not a yes or no question, but asking what user you want use (and only showing what the current default would be if you hit enter).

Also, account was spelled wrong.