Fixes #19205 - Add IPMI IP scanning to BMC API
This commit adds a new feature to smart-proxy's bmc module. This new
feature is an IP address range scanner that returns a list of IP
addresses that respond to IPMI ping packets on UDP port 623.

The use case for this is the discovery of IPMI hosts. Once the API
method has returned the list of responding IPMI hosts, another
application (perhaps Foreman in the future) can iterate over the list to
add the associated servers into Foreman as Foreman hosts and/or
configure the IPMI hosts in existing Foreman hosts' interfaces.

Scanning for IPMI hosts would be an important first step to onboard new
servers into Foreman if the current data center infrastructure
management does not yet have a reliable source of truth of which servers
have which IPMI hosts. Knowing which IP addresses are IPMI hosts would
simplify finding servers and IPMI NICs that are not in Foreman.

New API methods:

 - GET /bmc/scan
   Shows available resources /scan/range and /scan/cidr

 - GET /bmc/scan/range
   Shows usage on specifying a beginning IP address and an ending IP
   address for making a scan request

 - GET /bmc/scan/cidr
   Shows usage on specifying an IP address and its netmask in dot
   decimal format or prefixlen format for making a scan request

 - GET /bmc/scan/range/:address_first/:address_last
   Performs an IPMI ping scan from :address_first to :address_last and
   returns the result in key "result" of a JSON hash

 - GET /bmc/scan/cidr/:address/:netmask
   Performs an IPMI ping scan in the CIDR range of :address/:netmask,
   where :netmask is in decimal format (e.g. "255.255.255.0") or in
   prefixlen format (e.g. "24")

New classes:

 - Proxy::BMC::BaseScanner
   The interface for IP range scanning

 - Proxy::BMC::IPMIScanner
   An IP range scanner for IPMI. Implements Proxy::BMC::BaseScanner.
   Loosely based on ipmiutil_discover: http://ipmiutil.sourceforge.net/
   This class is used when the bmc_provider or bmc_default_provider is
   set to "freeipmi" or "ipmitool" because those imply IPMI BMC.

New configurable items in "bmc.yml":

 - :bmc_scanner_max_range_size
   The largest number of IP addresses that may be scanned in one
   request. Set this value to 0 to turn off the scanner.  Defaults to
   65536, which covers a CIDR prefixlen of /16.

 - :bmc_scanner_max_threads_per_request
   The maximum number of IP addresses to scan at the same time, per
   request. Defaults to 500.  Before increasing this number, you should
   increase the system limit on the maximum number of open files for the
   Smart Proxy process.

 - :bmc_scanner_socket_timeout_seconds
   How many seconds to wait for each scanned IP address to respond
   before timing out. Defaults to 1.