Fixes #20371 - Add sshfp DNS record support
as per

    https://tools.ietf.org/html/rfc4255
    https://tools.ietf.org/html/rfc6594

looks like

    $ curl -d 'fqdn=testsshfp.example.com&value=3 2 1235&type=SSHFP' \
        http://localhost:8000/dns
    $ host -t SSHFP testsshfp.example.com
    testsshfp.example.com has SSHFP record 3 2 1235
    $ curl -X DELETE http://localhost:8000/dns/testsshfp.example.com/SSHFP
    $ host -t SSHFP testsshfp.example.com
    Host testsshfp.example.com not found: 3(NXDOMAIN)

Due to the current smart-proxy API this is a bit unsymmetric.  While we
can create individual records we can only delete all of them.

That's fine for my use case: Letting a Foreman plugin handle SSH-keys
via SSHFP records so one gets rid of "host key has changed prompts once
and for all" (making ssh keys part of the VM life cycle) - but we might
want to extend this at a later point.

In order to not wait for

    https://github.com/rubysl/rubysl-resolv/pull/1

I've added the SSHFP record type to modules/dns_common/dns_resources.rb
so we can add other missing record types as well if missing.