Fixes #29463 - Drop webrick patch
In 508a9e8b18d8811f65f0f9f4201c287bae55da2d a webrick patch was added to
support setting ciphers. This is a copy of the built in method which
means it gets out of date. Ruby 2.4 added the SSLCiphers option so this
workaround is no longer needed.

As an example, the built in Ruby has correctly updated the automatic
self signed certificate to 2048 bits. It also supports Server Name
Indication.