Fixes #30856 - Drop  Puppet class cache initializer
Retrieving all classes for all environments only makes sense if caching
is enabled on the Puppetserver. By default this is the case. Even if
this is not the case, it only works for environments that don't change.
It's sufficient to lazy load this.

This saves load on start up of services.