Refs #30677 - support also urlencoded data
The default registration template in Foreman started to send the data in
urlencoded form instead of a JSON, as of https://projects.theforeman.org/issues/31043

In order to support that, we need to remove nested parameters from query,
since Sinatra parses it from body automatically. We only need to proxy
the body of the request. ForemanRequestFactory also defaults to
application/json if not set explicitly, in this case we need to enforce
the correct content-type through headers.

This also requires a change in handling headers, because if content-type
is specified as a header, it's set twice. Once it's set by
content_type=, second time through add_field, resulting in broken
header.