Fixes #31537 - Expose BMC providers as capabilities
Currently the BMC module exposes an API endpoint /providers/installed
but this can also be provided as capabilities or settings. This allows
Foreman to determine up front if a proxy is even a valid choice without
making API requests over the network.

The load_classes part is needed because capabilities are exectuted
before loading the main API (which already has the same require). The
require is explicit in the tests to stub out the implementation, making
the tests reliable, regardless of which providers are installed.