Fixes #35184 - Drop puppetca_puppet_cert provider
Currently for Puppet 6.0 or higher the Puppetserver REST API is used.
Puppet 5 is EOL so this can be dropped. Older versions (dating back to
Puppet 4) had an incomplete API, but core methods do work. Just the
reporting in the UI is incomplete. This means even with Puppet 5 it
won't be completely broken.

Dropping this allows us to move forward. The current implementation has
some coding techniques which need adjustment on Ruby 3. Since nobody
should be using the old code, refactoring it is pointless.