Wait for TFTP downloads
The /fetch_boot_file API is used to download TFTP files. The
HttpDownload task is implemented as a thread. Previously it was fired
off and there was no way to see the result. This implements code that
waits for some time before telling the client the task is still running.
There is no way for the client to find out if it really did finish.